[refactor] 관심사 분리를 위한 폴더 구조 변경 및 서비스 분리
