--- src/main/java/org/hana/wooahhanaapi/domain/account/adapter/AccountTransferAdapter.java
@@ -1,48 +1,0 @@
-package org.hana.wooahhanaapi.domain.account.adapter;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import org.hana.wooahhanaapi.domain.account.adapter.dto.AccountTransferReqDto;
-import org.hana.wooahhanaapi.domain.account.adapter.dto.AccountTransferRespDto;
-import org.hana.wooahhanaapi.domain.account.exception.TransferNotValidException;
-import org.springframework.stereotype.Service;
-
-import java.net.URI;
-import java.net.http.HttpClient;
-import java.net.http.HttpRequest;
-import java.net.http.HttpResponse;
-
-@Service
-public class AccountTransferAdapter implements AccountTransferPort{
-    // HttpClient 객체를 클래스 수준에서 재사용
-    private static final HttpClient client = HttpClient.newHttpClient();
-    private static final ObjectMapper objectMapper = new ObjectMapper();
-
-
-    @Override
-    public AccountTransferRespDto createAccountTransfer(AccountTransferReqDto reqDto) {
-        try {
-            String jsonBody = objectMapper.writeValueAsString(reqDto);
-
-            HttpRequest httpRequest = HttpRequest.newBuilder()
-                    .uri(URI.create("http://localhost:8000/transfer"))
-                    .version(HttpClient.Version.HTTP_1_1)
-                    .header("Content-Type", "application/json") // JSON 형식으로 전송
-                    .POST(HttpRequest.BodyPublishers.ofString(jsonBody))
-                    .build();
-
-            HttpResponse<String> response = client.send(httpRequest, HttpResponse.BodyHandlers.ofString());
-            System.out.println(response.statusCode());
-            System.out.println(response.body());
-            if(response.statusCode() == 400 || response.body().contains("(입출금) 잘못된 접근입니다.")) {
-                throw new TransferNotValidException("(입출금) 잘못된 접근입니다.");
-            }
-            else {
-                return objectMapper.readValue(response.body(), AccountTransferRespDto.class);
-            }
-
-        }
-        catch (Exception e){
-            throw new RuntimeException("Error occurred while fetching transfer records", e);
-        }
-    }
-}
