+++ src/main/java/org/hana/wooahhanaapi/domain/community/adapter/SaveValidCodePort.java
@@ -1,0 +1,5 @@
+package org.hana.wooahhanaapi.domain.community.adapter;
+
+public interface SaveValidCodePort {
+    void saveValidCode(String accountNumber, String validCode);
+}
