--- src/main/java/org/hana/wooahhanaapi/domain/community/adapter/SendValidCodePort.java
@@ -1,7 +1,0 @@
-package org.hana.wooahhanaapi.domain.community.adapter;
-
-import org.hana.wooahhanaapi.domain.community.adapter.dto.SendValidationCodeReqDto;
-
-public interface SendValidCodePort {
-    void sendValidCode(SendValidationCodeReqDto reqDto);
-}
